refactor(primitives): deprecate the legacy file module in place
Module-level deprecation inherits to every contained item, so entry_ref
is rehoused at the crate root: the width-agnostic reference union
outlives the legacy pipeline. Consumers still on the legacy surface
carry targeted allows until their nectar-file bridges land; the crate
root re-exports keep resolving.
