+++ crates/mantaray/src/error.rs
@@ -141,6 +141,7 @@ pub enum MantarayError {
     #[error(transparent)]
     Primitives(#[from] PrimitivesError),
     /// Error from the file splitter or joiner across the file/manifest seam.
+    #[allow(deprecated)]
     #[error(transparent)]
     File(#[from] nectar_primitives::file::FileError),
     /// Error from the typed chunk store during get operations.
